moved ProductCard.js from pages folder to components folder
